@@ -3,49 +3,42 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.15",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/styles": "^4.11.4",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@mui/icons-material": "^5.0.0",
     "@mui/material": "^5.2.8",
     "@stripe/react-stripe-js": "^1.5.0",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "@types/bcrypt": "^5.0.0",
-    "@types/crypto-js": "^4.0.2",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^12.20.25",
-    "@types/react": "^17.0.21",
-    "@types/react-dom": "^17.0.9",
-    "clsx": "^1.1.1",
-    "connected-react-router": "^6.9.1",
+    "clsx": "^1.2.1",
+    "connected-react-router": "^6.9.3",
     "crypto-js": "^4.1.1",
     "gsap": "^3.8.0",
     "history": "^4.10.1",
     "lodash": "^4.17.21",
     "notistack": "^1.0.10",
     "prettier": "^2.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-google-login": "^5.2.2",
     "react-loader-spinner": "^4.0.0",
     "react-redux": "^7.2.6",
     "react-router": "^5.2.1",
     "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^5.0.1",
     "react-stripe-checkout": "^2.6.3",
     "redux": "^4.1.2",
     "redux-persist": "^6.0.0",
     "redux-saga": "^1.1.3",
+    "tss-react": "^4.5.2",
     "typescript": "^4.4.3",
-    "use-debounce": "^7.0.0",
+    "use-debounce": "^9.0.3",
     "web-vitals": "^1.1.2",
-    "webp-loader": "^0.6.0"
+    "webp-loader": "^0.4.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -72,8 +65,14 @@
     ]
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.0",
+    "@types/crypto-js": "^4.1.1",
     "@types/history": "^4.7.9",
+    "@types/jest": "^29.2.6",
     "@types/lodash": "^4.14.173",
+    "@types/node": "^18.11.18",
+    "@types/react": "^17.0.53",
+    "@types/react-dom": "^18.0.10",
     "@types/react-router-dom": "^5.1.9"
   }
 }
