@@ -1,7 +1,7 @@
-import { makeStyles } from '@material-ui/styles';
+import { makeStyles } from 'tss-react/mui';
 
 
-export const useStyles = makeStyles((theme) => ({
+export const useStyles = makeStyles()((theme) => ({
     container: {
         display: 'flex',
         alignItems: 'center',
