@@ -1,4 +1,3 @@
-import { Paper } from '@material-ui/core';
 import React, { useCallback } from 'react';
 import Books from "../../assets/books.jpg";
 import Navbar from '../../components/Navbar';
@@ -8,7 +7,7 @@ import { useDispatch } from 'react-redux';
 import { setCreateClient } from '../../store/actions/auth.actions';
 
 const SignUpPage = () => {
-    const classes = useStyles();
+    const { classes } = useStyles();
     const dispatch = useDispatch();
 
     const signUp = ({ password:pass, email }:any) => {
